Change ingredient data type to Double to match backend

+++ app/src/main/java/com/scouts/kitchenplaner/model/entities/RecipeForCooking.kt
@@ -43,7 +43,7 @@ class RecipeForCooking(
             it.ingredients.map { ingredient ->
                 Ingredient(
                     ingredient.name,
-                    (ingredient.amount * people / recipe.numberOfPeople.toDouble()).toFloat(),
+                    (ingredient.amount * people / recipe.numberOfPeople.toDouble()),
                     ingredient.unit
                 )
             }
